// frontend/src/app/core/models/sort-enum.ts
export enum SortEnum {
    default = "Newest",
    luckAndLevel = "Highest Level",
    attack = "Highest Attack",
    magic = "Highest Magic",
    shield = "Highest Shield",
    feathers = "Highest Feathers",
    team_a = "Team A First",
    team_b = "Team B First",
    droppings = "Least Droppings"
}